Remove error type
